% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHBXII}
\alias{CHBXII}
\title{Burr-XII  (BXII) Cumulative Hazard Function.}
\usage{
CHBXII(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}
}
\value{
the value of the BXII cumulative hazard function
}
\description{
Burr-XII  (BXII) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHBXII(t=t, kappa=0.5, alpha=0.35)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
