% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHG}
\alias{CHG}
\title{Gamma (G) Cumulative Hazard Function.}
\usage{
CHG(t, shape, scale)
}
\arguments{
\item{t}{: positive argument}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}
}
\value{
the value of the G cumulative hazard function
}
\description{
Gamma (G) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHG(t=t, shape=0.85, scale=0.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
