% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHGG}
\alias{CHGG}
\title{Generalised Gamma (GG) Cumulative Hazard Function.}
\usage{
CHGG(t, kappa, alpha, eta)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}
}
\value{
the value of the GG cumulative hazard function
}
\description{
Generalised Gamma (GG) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHGG(t=t, kappa=0.5, alpha=0.35, eta=0.9)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
