% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHKW}
\alias{CHKW}
\title{Kumaraswamy Weibull (KW) Cumulative Hazard Function.}
\usage{
CHKW(t, alpha, kappa, eta, zeta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{kappa}{: shape parameter}

\item{eta}{: shape parameter}

\item{zeta}{: shape parameter}
}
\value{
the value of the KW cumulative hazard function
}
\description{
Kumaraswamy Weibull (KW) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHKW(t=t, alpha=0.35, kappa=0.5, eta=1.20, zeta=1.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
