% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHLL}
\alias{CHLL}
\title{Log-logistic  (LL) Cumulative Hazard Function.}
\usage{
CHLL(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}
}
\value{
the value of the LL cumulative hazard function
}
\description{
Log-logistic  (LL) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHLL(t=t, kappa=0.5, alpha=0.35)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
