% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHLN}
\alias{CHLN}
\title{Lognormal (LN) Cumulative Hazard Function.}
\usage{
CHLN(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: meanlog parameter}

\item{alpha}{: sdlog parameter}
}
\value{
the value of the LN cumulative hazard function
}
\description{
Lognormal (LN) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHLN(t=t, kappa=0.75, alpha=0.95)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
