% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHMKW}
\alias{CHMKW}
\title{Modified Kumaraswamy Weibull (MKW) Cumulative Hazard Function.}
\usage{
CHMKW(t, alpha, kappa, eta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: Inverse scale parameter}

\item{kappa}{: shape parameter}

\item{eta}{: shape parameter}
}
\value{
the value of the MKW cumulative hazard function
}
\description{
Modified Kumaraswamy Weibull (MKW) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHMKW(t=t,alpha=0.35, kappa=0.7, eta=1.4)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
