% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHNGLL}
\alias{CHNGLL}
\title{New Generalized Log-logistic (GLL) cumulative hazard function.}
\usage{
CHNGLL(t, kappa, alpha, eta, zeta)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{zeta}{: shape parameter}
}
\value{
the value of the NGLL cumulative hazard function
}
\description{
New Generalized Log-logistic (GLL) cumulative hazard function.
}
\examples{
t=runif(10,min=0,max=1)
CHNGLL(t=t, kappa=0.5, alpha=0.35, eta=0.7, zeta=1.4)

}
\references{
Hassan Muse, A. A new generalized log-logistic distribution with increasing, decreasing, unimodal and bathtub-shaped hazard rates: properties and applications, in Proceedings of the Symmetry 2021 - The 3rd International Conference on Symmetry, 8–13 August 2021, MDPI: Basel, Switzerland, doi:10.3390/Symmetry2021-10765.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
