% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHPGW}
\alias{CHPGW}
\title{Power Generalised Weibull (PGW) cumulative hazard function.}
\usage{
CHPGW(t, kappa, alpha, eta)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}
}
\value{
the value of the PGW cumulative hazard function
}
\description{
Power Generalised Weibull (PGW) cumulative hazard function.
}
\examples{
t=runif(10,min=0,max=1)
CHPGW(t=t, kappa=0.5, alpha=1.5, eta=0.6)

}
\references{
Alvares, D., & Rubio, F. J. (2021). A tractable Bayesian joint model for longitudinal and survival data. Statistics in Medicine, 40(19), 4213-4229.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
