% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{CHW}
\alias{CHW}
\title{Weibull  (W) Cumulative Hazard Function.}
\usage{
CHW(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}
}
\value{
the value of the W cumulative hazard function
}
\description{
Weibull  (W) Cumulative Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
CHW(t=t, kappa=0.75, alpha=0.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
