% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{dggamma}
\alias{dggamma}
\title{Generalised Gamma (GG) Probability Density Function.}
\usage{
dggamma(t, kappa, alpha, eta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the GG probability density function
}
\description{
Generalised Gamma (GG) Probability Density Function.
}
\examples{
t=runif(10,min=0,max=1)
dggamma(t=t, kappa=0.5, alpha=0.35, eta=0.9,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
