% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1684.R
\docType{data}
\name{e1684}
\alias{e1684}
\title{Melanoma  data set}
\format{
A data frame with 284 observations on the following 5 variables.
 \itemize{
  \item TRT: 0=control group, 1=IFN treatment group
   \item FAILTIME: observed relapse-free time
   \item FAILCENS: relapse-free censor indicator
  \item AGE:continuous variable, which is centered to the mean
   \item SEX: 0 for male, 1 fopr female
 }
}
\description{
Eastern Cooperative Oncology Group (ECOG) data used for modeling hazard-based regression models
}
\references{
Kirkwood, J. M., Manola, J., Ibrahim, J., Sondak, V., Ernstoff, M. S., & Rao, U. (2004). A pooled analysis of eastern cooperative oncology group and intergroup trials of adjuvant high-dose interferon for melanoma. Clinical Cancer Research, 10(5), 1670-1677.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
\keyword{datasets}
