% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hEW}
\alias{hEW}
\title{Exponentiated Weibull (EW) Hazard Function.}
\usage{
hEW(t, lambda, kappa, alpha, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{lambda}{: scale parameter}

\item{kappa}{: shape parameter}

\item{alpha}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the EW hazard function
}
\description{
Exponentiated Weibull (EW) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hEW(t=t, lambda=0.9, kappa=0.5, alpha=0.75, log=FALSE)

}
\references{
Khan, S. A. (2018). Exponentiated Weibull regression for time-to-event data. Lifetime data analysis, 24(2), 328-354.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
