% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hGG}
\alias{hGG}
\title{Generalised Gamma (GG) Hazard Function.}
\usage{
hGG(t, kappa, alpha, eta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the GG hazard function
}
\description{
Generalised Gamma (GG) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hGG(t=t, kappa=0.5, alpha=0.35, eta=0.9,log=FALSE)

}
\references{
Agarwal, S. K., & Kalla, S. L. (1996). A generalized gamma distribution and its application in reliabilty. Communications in Statistics-Theory and Methods, 25(1), 201-210.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
