% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hKW}
\alias{hKW}
\title{Kumaraswamy Weibull (KW) Hazard Function.}
\usage{
hKW(t, alpha, kappa, eta, zeta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{kappa}{: shape parameter}

\item{eta}{: shape parameter}

\item{zeta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the KW hazard function
}
\description{
Kumaraswamy Weibull (KW) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hKW(t=t, alpha=0.35, kappa=0.5, eta=1.20, zeta=1.5, log=FALSE)

}
\references{
Cordeiro, G. M., Ortega, E. M., & Nadarajah, S. (2010). The Kumaraswamy Weibull distribution with application to failure data. Journal of the Franklin Institute, 347(8), 1399-1429.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
