% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hLL}
\alias{hLL}
\title{Log-logistic (LL) Hazard Function.}
\usage{
hLL(t, kappa, alpha, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the LL hazard function
}
\description{
Log-logistic (LL) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hLL(t=t, kappa=0.5, alpha=0.35,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
