% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hLN}
\alias{hLN}
\title{Lognormal (LN) Hazard Function.}
\usage{
hLN(t, kappa, alpha, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: meanlog parameter}

\item{alpha}{: sdlog parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the LN hazard function
}
\description{
Lognormal (LN) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hLN(t=t, kappa=0.5, alpha=0.75,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
