% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hPGW}
\alias{hPGW}
\title{Power Generalised Weibull (PGW) hazard function.}
\usage{
hPGW(t, kappa, alpha, eta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the PGW hazard function
}
\description{
Power Generalised Weibull (PGW) hazard function.
}
\examples{
t=runif(10,min=0,max=1)
hPGW(t=t, kappa=0.5, alpha=1.5, eta=0.6,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
