% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{hW}
\alias{hW}
\title{Weibull (W) Hazard Function.}
\usage{
hW(t, kappa, alpha, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the w hazard function
}
\description{
Weibull (W) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
hW(t=t, kappa=0.75, alpha=0.5,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
