% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHReg1.R
\name{sggamma}
\alias{sggamma}
\title{Generalised Gamma (GG) Survival Function.}
\usage{
sggamma(t, kappa, alpha, eta, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{log.p}{:log scale (TRUE or FALSE)}
}
\value{
the value of the GG survival function
}
\description{
Generalised Gamma (GG) Survival Function.
}
\examples{
t=runif(10,min=0,max=1)
sggamma(t=t, kappa=0.5, alpha=0.35, eta=0.9,log.p=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Mutua Kilai, \email{abdisalam.hassan@amoud.edu.so}
}
